% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{aflp A data frame with 100 observations on the following 31 variables.}
\alias{aflp A data frame with 100 observations on the following 31 variables.}
\title{AFLP dataset}
\format{a list with two elements}
\description{
An example dataset including a phylogeny of 13 Galapagos finches with
trait data matched to the phylogeny.
}
\details{
\describe{
 \item{pop}{phylogeny of Galapagos finches}
 \item{data}{measured traits for each species in the phylogeny}
}
}
\keyword{datasets}

