% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{aflp AFLP marker dataset}
\alias{aflp AFLP marker dataset}
\title{AFLP dataset}
\format{A data frame with 100 observations on the following 31 variables.}
\description{
An example dataset including a phylogeny of 13 Galapagos finches with
trait data matched to the phylogeny.
}
\details{
\describe{
 \item{pop}{phylogeny of Galapagos finches}
 \item{q_hat}{columns 2 through 16 are the q_hat values for each locus in each population}
	\item{q_var}{columns 17 through 31 are the q_hat variances for each locus in each population}
}
}
\keyword{datasets}

