\name{biallelic}
\alias{biallelic}
\docType{data}
\title{
Biallelic marker dataset
}
\description{
An example dataset created from simulations of 15 neutral populations under an island model. Individuals are diploid with 25 biallelic loci.
}
\usage{data(biallelic)}
\format{
  A data frame with 375 observations on the following 26 variables.
  \describe{
    \item{\code{pop}}{a numeric vector}
    \item{\code{loc1}}{a numeric vector}
    \item{\code{loc2}}{a numeric vector}
    \item{\code{loc3}}{a numeric vector}
    \item{\code{loc4}}{a numeric vector}
    \item{\code{loc5}}{a numeric vector}
    \item{\code{loc6}}{a numeric vector}
    \item{\code{loc7}}{a numeric vector}
    \item{\code{loc8}}{a numeric vector}
    \item{\code{loc9}}{a numeric vector}
    \item{\code{loc10}}{a numeric vector}
    \item{\code{loc11}}{a numeric vector}
    \item{\code{loc12}}{a numeric vector}
    \item{\code{loc13}}{a numeric vector}
    \item{\code{loc14}}{a numeric vector}
    \item{\code{loc15}}{a numeric vector}
    \item{\code{loc16}}{a numeric vector}
    \item{\code{loc17}}{a numeric vector}
    \item{\code{loc18}}{a numeric vector}
    \item{\code{loc19}}{a numeric vector}
    \item{\code{loc20}}{a numeric vector}
    \item{\code{loc21}}{a numeric vector}
    \item{\code{loc22}}{a numeric vector}
    \item{\code{loc23}}{a numeric vector}
    \item{\code{loc24}}{a numeric vector}
    \item{\code{loc25}}{a numeric vector}
  }
}
\details{
	\describe{
		\item{pop}{population of origin}
		\item{loc}{loc1 through loc25 are the 25 biallelic loci for which each diploid individual is genotyped}
	}
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(biallelic)
## maybe str(biallelic) ; plot(biallelic) ...
}
\keyword{datasets}
