% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/QstFstComp.R
\docType{data}
\name{hsdam}
\alias{hsdam}
\title{Half-sib dam trait dataset}
\format{A data frame with 900 observations on the following 3 variables.}
\description{
An example dataset of a neutral trait created from simulations of 15
populations under an island model. Offspring per dam are unbalanced across
the breeding design with 6 dams per population, an average of 10 offspring per
dam, and 900 individuals total.
}
\details{
\describe{
 \item{pop}{a column of population identifiers}
 \item{dam}{a column of unique dam identifiers}
 \item{trait}{column of trait values per individual}
}
}
\keyword{datasets}

