% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{hssire}
\alias{hssire}
\title{Half-sib sire trait dataset}
\format{A data frame with 900 observations on the following 4 variables.}
\description{
An example dataset of a neutral trait created from simulations of 15 populations
under an island model. 900 individuals total come from an average of 4 dams
per sire and an average of 12 sires per population for an overall average of 60
individuals per population.
}
\details{
\describe{
 \item{pop}{a column of population identifiers}
 \item{sire}{a column of unique sire identifiers}
 \item{dam}{a column of unique dam identifiers}
 \item{trait}{column of trait values per individual}
}
}
\keyword{datasets}

