% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QstFstComp.R
\docType{data}
\name{biallelic}
\alias{biallelic}
\title{Biallelic marker dataset}
\format{A data frame with 375 observations on the following 26 variables.}
\description{
An example dataset created from simulations of 15 neutral populations under 
an island model. Individuals are diploid with genotypes at 25 biallelic loci.
}
\details{
\describe{
 \item{pop}{population of origin}
 \item{loc}{loc1 through loc25 are the 25 biallelic loci for which each diploid individual is genotyped}
}
}
\keyword{datasets}
